requirements.txt
pandas==1.3.4
pandas_ta==0.3.14b0
yfinance==0.2.31
python-dotenv==1.0.0
matplotlib==3.4.3
matplotlib-inline==0.1.2
feed-parser==6.0.11